# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * NY Property Tax Grievance Assistant API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
  message: string;
}

export type GrievanceStatus =
  (typeof GrievanceStatus)[keyof typeof GrievanceStatus];

export const GrievanceStatus = {
  draft: "draft",
  submitted: "submitted",
  pending: "pending",
  reduced: "reduced",
  denied: "denied",
} as const;

export interface Grievance {
  id: number;
  ownerName: string;
  propertyAddress: string;
  county: string;
  municipality: string;
  taxYear: number;
  currentAssessment: number;
  equalizationRate?: number;
  estimatedMarketValue: number;
  requestedAssessment: number;
  status: GrievanceStatus;
  filingDeadline?: string | null;
  notes?: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface CreateGrievanceRequest {
  ownerName: string;
  propertyAddress: string;
  county: string;
  municipality: string;
  taxYear: number;
  currentAssessment: number;
  equalizationRate?: number;
  estimatedMarketValue: number;
  requestedAssessment: number;
  filingDeadline?: string | null;
  notes?: string | null;
}

export type UpdateGrievanceRequestStatus =
  (typeof UpdateGrievanceRequestStatus)[keyof typeof UpdateGrievanceRequestStatus];

export const UpdateGrievanceRequestStatus = {
  draft: "draft",
  submitted: "submitted",
  pending: "pending",
  reduced: "reduced",
  denied: "denied",
} as const;

export interface UpdateGrievanceRequest {
  ownerName?: string;
  propertyAddress?: string;
  county?: string;
  municipality?: string;
  taxYear?: number;
  currentAssessment?: number;
  equalizationRate?: number;
  estimatedMarketValue?: number;
  requestedAssessment?: number;
  status?: UpdateGrievanceRequestStatus;
  filingDeadline?: string | null;
  notes?: string | null;
}

export interface Comparable {
  id: number;
  grievanceId: number;
  address: string;
  salePrice: number;
  saleDate: string;
  squareFeet?: number | null;
  bedrooms?: number | null;
  bathrooms?: number | null;
  assessedValue?: number | null;
  notes?: string | null;
  createdAt: string;
}

export interface CreateComparableRequest {
  grievanceId: number;
  address: string;
  salePrice: number;
  saleDate: string;
  squareFeet?: number | null;
  bedrooms?: number | null;
  bathrooms?: number | null;
  assessedValue?: number | null;
  notes?: string | null;
}

export interface County {
  id: string;
  name: string;
  region: string;
  filingPortal?: string | null;
  formRequired: string;
  grievanceDay?: string | null;
  deadline?: string | null;
  notes: string;
  municipalities: string[];
}

export type ListComparablesParams = {
  grievanceId: number;
};