# lib/api-zod/src/generated/types/comparable.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * NY Property Tax Grievance Assistant API
 * OpenAPI spec version: 0.1.0
 */

export interface Comparable {
  id: number;
  grievanceId: number;
  address: string;
  salePrice: number;
  saleDate: string;
  squareFeet?: number | null;
  bedrooms?: number | null;
  bathrooms?: number | null;
  assessedValue?: number | null;
  notes?: string | null;
  createdAt: string;
}