# lib/api-zod/src/generated/types/county.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * NY Property Tax Grievance Assistant API
 * OpenAPI spec version: 0.1.0
 */

export interface County {
  id: string;
  name: string;
  region: string;
  filingPortal?: string | null;
  formRequired: string;
  grievanceDay?: string | null;
  deadline?: string | null;
  notes: string;
  municipalities: string[];
}