# lib/api-zod/src/generated/types/createGrievanceRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * NY Property Tax Grievance Assistant API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateGrievanceRequest {
  ownerName: string;
  propertyAddress: string;
  county: string;
  municipality: string;
  taxYear: number;
  currentAssessment: number;
  equalizationRate?: number;
  estimatedMarketValue: number;
  requestedAssessment: number;
  filingDeadline?: string | null;
  notes?: string | null;
}