# lib/api-zod/src/generated/types/updateGrievanceRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * NY Property Tax Grievance Assistant API
 * OpenAPI spec version: 0.1.0
 */
import type { UpdateGrievanceRequestStatus } from "./updateGrievanceRequestStatus";

export interface UpdateGrievanceRequest {
  ownerName?: string;
  propertyAddress?: string;
  county?: string;
  municipality?: string;
  taxYear?: number;
  currentAssessment?: number;
  equalizationRate?: number;
  estimatedMarketValue?: number;
  requestedAssessment?: number;
  status?: UpdateGrievanceRequestStatus;
  filingDeadline?: string | null;
  notes?: string | null;
}